+++ crates/vmi-arch-amd64/src/lib.rs
@@ -25,9 +25,8 @@ pub use self::{
     dr::{Dr0, Dr1, Dr2, Dr3, Dr6, Dr7},
     efer::MsrEfer,
     event::{
-        EventCpuId, EventGuestRequest, EventInterrupt, EventIo, EventIoDirection,
-        EventMemoryAccess, EventMonitor, EventReason, EventSinglestep, EventWriteControlRegister,
-        MemoryAccessFlags,
+        EventCpuId, EventHypercall, EventInterrupt, EventIo, EventIoDirection, EventMemoryAccess,
+        EventMonitor, EventReason, EventSinglestep, EventWriteControlRegister, MemoryAccessFlags,
     },
     interrupt::{ExceptionVector, Idt, IdtAccess, IdtEntry, Interrupt, InterruptType},
     paging::{PageTableEntry, PageTableLevel, PagingMode},