+++ crates/vmi-utils/src/injector/os/windows/kernel_mode.rs
@@ -179,7 +179,7 @@ where
             EventReason::MemoryAccess(_) => self.on_memory_access(vmi),
             EventReason::Interrupt(_) => self.on_interrupt(vmi),
             EventReason::Singlestep(_) => self.on_singlestep(vmi),
-            EventReason::GuestRequest(_) => self.on_vmcall(vmi),
+            EventReason::Hypercall(_) => self.on_hypercall(vmi),
             _ => panic!("Unhandled event: {:?}", vmi.event().reason()),
         }
     }
@@ -364,7 +364,7 @@ where
             }
             else {
                 self.state = InjectorState::Bridge;
-                vmi.monitor_enable(EventMonitor::GuestRequest {
+                vmi.monitor_enable(EventMonitor::Hypercall {
                     allow_userspace: false,
                 })?;
             }
@@ -381,15 +381,15 @@ where
         Ok(VmiEventResponse::default().with_view(self.view))
     }
 
-    #[tracing::instrument(name = "vmcall", skip_all, err)]
-    fn on_vmcall(
+    #[tracing::instrument(name = "hypercall", skip_all, err)]
+    fn on_hypercall(
         &mut self,
         vmi: &VmiContext<WindowsOs<Driver>>,
     ) -> Result<VmiEventResponse<Amd64>, VmiError> {
-        let guest_request = vmi.event().reason().as_guest_request();
+        let hypercall = vmi.event().reason().as_hypercall();
 
         let mut registers = vmi.registers().gp_registers();
-        registers.rip += guest_request.instruction_length as u64;
+        registers.rip += hypercall.instruction_length as u64;
 
         tracing::trace!(
             magic = %Hex(registers.rbp as u32),
@@ -415,7 +415,7 @@ where
 
             if let Some(complete) = complete {
                 self.state = InjectorState::Complete(complete);
-                vmi.monitor_disable(EventMonitor::GuestRequest {
+                vmi.monitor_disable(EventMonitor::Hypercall {
                     allow_userspace: false,
                 })?;
             }
@@ -460,14 +460,14 @@ where
         // Disabled when transitioning from `Teardown` to `Bridge` or `Complete`.
         let mut disable_interrupt = false;
         // Disabled when transitioning from `Bridge` to `Complete`.
-        let mut disable_guest_request = false;
+        let mut disable_hypercall = false;
 
         match self.state {
             InjectorState::PreHijack | InjectorState::Executing | InjectorState::Teardown(_) => {
                 disable_interrupt = true;
             }
             InjectorState::Bridge => {
-                disable_guest_request = true;
+                disable_hypercall = true;
             }
             _ => {}
         }
@@ -492,14 +492,14 @@ where
             }
         }
 
-        // In `Bridge` state, we are guaranteed to have enabled the guest request
+        // In `Bridge` state, we are guaranteed to have enabled the hypercall
         // monitor, so we must disable it.
-        if disable_guest_request
-            && let Err(err) = vmi.monitor_disable(EventMonitor::GuestRequest {
+        if disable_hypercall
+            && let Err(err) = vmi.monitor_disable(EventMonitor::Hypercall {
                 allow_userspace: false,
             })
         {
-            tracing::error!(%err, "failed to disable guest request monitor");
+            tracing::error!(%err, "failed to disable hypercall monitor");
         }
 
         // In all states, we are guaranteed to have enabled the singlestep